Quite a bit of debugging here, as noticed when graphing with exact values instead of averages, total energy in the system isnt behaving as it should, and this was all due to using float instead of double... Also i noticed that bonds were vibrating way too much, and realised that i was using the wrong k values for things like bonds and angles, as the ITP files use the func 2, which change the values depending on the equations you use, so that made things at the step-level much better and not so crazy. However, all the physics falls apart when new lists get built, and only noticed now as i had incorrect values for cutoffs, so new lists were not being built as they should. That is a fix for tomorrow.

+++ MD_Engine/cuda/bond_kernel.cu
@@ -1,14 +1,14 @@
 extern "C" __global__
 
 void bond_kernel(
-    const float *positions, // the * is reciving the adress of the array itself on the GPU instead of an actual array
-    float *forces,          // this makes the variable a pointer, pointing to the position in memory
+    const double *positions, // the * is reciving the adress of the array itself on the GPU instead of an actual array
+    double *forces,          // this makes the variable a pointer, pointing to the position in memory
     const int *bond_a,
     const int *bond_b,
-    const float *bond_r0,
-    const float *bond_k,
-    float box_length,
-    float *potential_energy,
+    const double *bond_r0,
+    const double *bond_k,
+    double box_length,
+    double *potential_energy,
     int n_bonds
 )
 
@@ -29,60 +29,60 @@ void bond_kernel(
     int a = bond_a[bond]; // gets the values for one specific bond
     int b = bond_b[bond];
 
-    float r0 = bond_r0[bond];
-    float k = bond_k[bond];
+    double r0 = bond_r0[bond];
+    double k = bond_k[bond];
 
     // -------------------------------------------------
     // 3. Read atom coordinates
     // The data is stored as a long list of numbers, and every three numbers is equiviant to the x,y and z values
     // -------------------------------------------------
 
-    float x1 = positions[3 * a + 0]; // gets the x pos value of atom a, reads it, and store in the register x1
-    float y1 = positions[3 * a + 1]; // gets the y pos value of atom a, reads it, and store in the register y1
-    float z1 = positions[3 * a + 2]; // gets the z pos value of atom a, reads it, and store in the register z1
+    double x1 = positions[3 * a + 0]; // gets the x pos value of atom a, reads it, and store in the register x1
+    double y1 = positions[3 * a + 1]; // gets the y pos value of atom a, reads it, and store in the register y1
+    double z1 = positions[3 * a + 2]; // gets the z pos value of atom a, reads it, and store in the register z1
 
-    float x2 = positions[3 * b + 0]; // gets the x pos value of atom b, reads it, and store in the register x2
-    float y2 = positions[3 * b + 1]; // gets the y pos value of atom b, reads it, and store in the register y2
-    float z2 = positions[3 * b + 2]; // gets the z pos value of atom b, reads it, and store in the register z2
+    double x2 = positions[3 * b + 0]; // gets the x pos value of atom b, reads it, and store in the register x2
+    double y2 = positions[3 * b + 1]; // gets the y pos value of atom b, reads it, and store in the register y2
+    double z2 = positions[3 * b + 2]; // gets the z pos value of atom b, reads it, and store in the register z2
 
     // -------------------------------------------------
     // 4. Bond vector - same as (system.positions[b] - system.positions[a])
     // -------------------------------------------------
 
-    float dx = x2 - x1;
-    float dy = y2 - y1;
-    float dz = z2 - z1;
+    double dx = x2 - x1;
+    double dy = y2 - y1;
+    double dz = z2 - z1;
 
     // -------------------------------------------------
     // 5. Minimum image - same as system.minimum_image()
     // -------------------------------------------------
     
-    dx -= box_length * roundf(dx / box_length);
-    dy -= box_length * roundf(dy / box_length);
-    dz -= box_length * roundf(dz / box_length);
+    dx -= box_length * round(dx / box_length);
+    dy -= box_length * round(dy / box_length);
+    dz -= box_length * round(dz / box_length);
 
     // -------------------------------------------------
     // 6. Distance - same as cp.linalg.norm(r_vec, axis=1)
     // -------------------------------------------------
 
-    float r2 = dx * dx + dy * dy + dz * dz;
+    double r2 = dx * dx + dy * dy + dz * dz;
 
     if (r2 < 1e-24f) // same as valid = r > 1e-12, without taking the sqrt
         return;
 
-    float r = sqrtf(r2);
+    double r = sqrt(r2);
 
-    float inv_r = 1.0f / r;
+    double inv_r = 1.0f / r;
 
     // -------------------------------------------------
     // 7. Harmonic bond force - same as F = 2 * k[:, None] * (r - r0)[:, None] * r_hat ...
     // -------------------------------------------------
 
-    float force_mag = 2.0f * k * (r - r0);
+    double force_mag = 2.0f * k * (r - r0);
 
-    float Fx = force_mag * dx * inv_r;
-    float Fy = force_mag * dy * inv_r;
-    float Fz = force_mag * dz * inv_r;
+    double Fx = force_mag * dx * inv_r;
+    double Fy = force_mag * dy * inv_r;
+    double Fz = force_mag * dz * inv_r;
 
     // -------------------------------------------------
     // 8. Apply Newton's Third Law - same as cp.add.at(system.forces, a,  F) ...
@@ -100,7 +100,7 @@ void bond_kernel(
     // 9. Bond potential energy - same as system.potential_energy += cp.sum(k * (r - r0)**2)
     // -------------------------------------------------
 
-    float energy = k * (r - r0) * (r - r0);
+    double energy = k * (r - r0) * (r - r0);
 
     atomicAdd(potential_energy, energy);
 }
